Clue system and serialization created.
