+++ Assets/Clues/Clue Database.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5dafb462e99818d4499e6607eb781f19, type: 3}
+  m_Name: Clue Database
+  m_EditorClassIdentifier: 
+  m_clues:
+  - {fileID: 11400000, guid: 4a1cf1ab969ce504690c601a61061052, type: 2}
+  - {fileID: 0}
