+++ Assets/Clues/Clue Database.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a24f39508c134c1498985c44200f745c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
