+++ Assets/Clues/Example Clue.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a1cf1ab969ce504690c601a61061052
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
