--- Assets/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c28d6c55ac69a294fad1ad93e087cd00
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
