--- Assets/Plugins.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aa71adf1b04026244a66a4287a6f579e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
