--- Assets/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0b89dce5f8833d3459e52bb6815faa63
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
