+++ Assets/Scripts/Clues/Clue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 772023b7c8de8994ab95ae2125e5c562
