+++ Assets/Scripts/Clues/ClueBoardClue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0d68056c767ee394abd522a401800a35
