+++ Assets/Scripts/Clues/ClueBoardNote.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7d5cae7ec050b3849b6d9334fb8b5a5f
