+++ Assets/Scripts/Clues/ClueDatabase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5dafb462e99818d4499e6607eb781f19
