+++ Assets/Scripts/GlobalManagers/ClueManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c9a0eba39e0e4234d92047203ca14cc4
