+++ Assets/Scripts/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86cf6338ad9d98842a6951cb732c1cd8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
