+++ Assets/Scripts/State/YarnStateLinker.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4ba087fe9885d63458b571231250790a
