--- Assets/State/Authored State.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 98e3a2a0dc912b94380975de78d9c9eb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
