+++ Assets/State/New State Asset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 781caaf5e01bfd840902d6040ddebb87
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
