--- Assets/State/State.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ac23c20cd090a174fab07b2e63976501
