--- Assets/State/StateAsset.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "StateAsset", menuName = "Scriptable Objects/State Asset")]
-public class StateAsset : ScriptableObject
-{
-    [SerializeField]
-    State m_state;
-    public State State => m_state;
-}
