--- Assets/State/StateAsset.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 237863071e5689f428ff3bec1d6ee3a2
