--- Assets/State/YarnStateLinker.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4ba087fe9885d63458b571231250790a
